+++ renderer/components/modals/ResultModal.tsx
@@ -3,7 +3,7 @@ import { Container, Sprite, Text } from '@inlet/react-pixi'
 import { InteractionEvent, TextStyle } from 'pixi.js'
 
 import { Position } from '../../types/character'
-import { containsPointClickThrouth } from '../../utils/pixi_api'
+import { containsPointClickThrouth } from '../../utils/PixiAPI'
 import Num from '../items/Num'
 import { shouldStrToNum } from '../../utils/api'
 import CloseBtn from '../buttons/CloseBtn'
