+++ renderer/pages/index.tsx
@@ -11,7 +11,7 @@ import LifeGauge from '../components/items/LifeGauge'
 import Coin from '../components/items/Coin'
 import CoinText from '../components/items/CoinText'
 import { Position } from '../types/character'
-import { containsPointClickThrouth } from '../utils/pixi_api'
+import { containsPointClickThrouth } from '../utils/PixiAPI'
 import StartBtn from '../components/buttons/StartBtn'
 import SettingBtn from '../components/buttons/SettingBtn'
 import EndBtn from '../components/buttons/EndBtn'
